Extract bootstrapping code out
