+++ www/src/js/bootstrapping/configure-store.js
